# example/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Projects API
  version: 1.0.0
  description: Manage projects.
servers:
  - url: https://api.example.com
security:
  - BearerAuth: []
paths:
  /v4/projects:
    post:
      operationId: createProject
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, ownerEmail]
              properties:
                name:
                  type: string
                ownerEmail:
                  type: string
                status:
                  $ref: '#/components/schemas/ProjectStatus'
      responses:
        '201':
          description: The created project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    get:
      operationId: listProjects
      summary: List projects
      parameters:
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProjectStatus'
        - name: limit
          in: query
          description: Page size, 1-100.
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: A page of projects.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Project'
                  nextCursor:
                    type: [string, 'null']
  /v4/projects/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
    get:
      operationId: getProject
      summary: Get a project
      responses:
        '200':
          description: The project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    patch:
      operationId: updateProject
      summary: Update a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                status:
                  $ref: '#/components/schemas/ProjectStatus'
                archived:
                  type: boolean
      responses:
        '200':
          description: The updated project.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
    delete:
      operationId: deleteProject
      summary: Delete a project
      responses:
        '204':
          description: The project was deleted.
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    ProjectStatus:
      type: string
      enum: [active, paused, completed]
    Project:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        ownerEmail:
          type: string
        status:
          $ref: '#/components/schemas/ProjectStatus'
        archived:
          type: boolean
        createdAt:
          type: string
          format: date-time